Make method interceptors general and rename them
