+++ tests/Go/Aop/Framework/AfterInterceptorTest.php
@@ -10,20 +10,20 @@
 
 namespace Go\Aop\Framework;
 
-class MethodAfterThrowingInterceptorTest extends AbstractMethodInterceptorTest
+class AfterInterceptorTest extends AbstractMethodInterceptorTest
 {
 
-    public function testAdviceIsNotCalledAfterInvocation()
+    public function testAdviceIsCalledAfterInvocation()
     {
         $sequence   = array();
         $advice     = $this->getAdvice($sequence);
         $invocation = $this->getInvocation($sequence);
 
-        $interceptor = new MethodAfterThrowingInterceptor($advice);
+        $interceptor = new AfterInterceptor($advice);
         $result = $interceptor->invoke($invocation);
 
         $this->assertEquals('invocation', $result, "Advice should not affect the return value of invocation");
-        $this->assertEquals(array('invocation'), $sequence, "Advice should not be invoked");
+        $this->assertEquals(array('invocation', 'advice'), $sequence, "After advice should be invoked after invocation");
     }
 
     public function testAdviceIsCalledAfterExceptionInInvocation()
@@ -32,13 +32,13 @@ class MethodAfterThrowingInterceptorTest extends AbstractMethodInterceptorTest
         $advice     = $this->getAdvice($sequence);
         $invocation = $this->getInvocation($sequence, true);
 
-        $interceptor = new MethodAfterThrowingInterceptor($advice);
+        $interceptor = new AfterInterceptor($advice);
         $this->setExpectedException('RuntimeException');
         try {
             $interceptor->invoke($invocation);
         } catch (\Exception $e) {
-            $this->assertEquals(array('invocation', 'advice'), $sequence, "Advice should be invoked after invocation");
+            $this->assertEquals(array('invocation', 'advice'), $sequence, "After advice should be invoked after invocation");
             throw $e;
         }
-    }}
- 
+    }
+}
