+++ tests/Go/Aop/Framework/AroundInterceptorTest.php
@@ -12,7 +12,7 @@ namespace Go\Aop\Framework;
 
 use Go\Aop\Intercept\MethodInvocation;
 
-class MethodAroundInterceptorTest extends AbstractMethodInterceptorTest
+class AroundInterceptorTest extends AbstractMethodInterceptorTest
 {
 
     public function testInvocationIsNotCalledWithoutProceed()
@@ -21,7 +21,7 @@ class MethodAroundInterceptorTest extends AbstractMethodInterceptorTest
         $advice     = $this->getAdvice($sequence); // advice will not call Invocation->proceed()
         $invocation = $this->getInvocation($sequence);
 
-        $interceptor = new MethodAroundInterceptor($advice);
+        $interceptor = new AroundInterceptor($advice);
         $result = $interceptor->invoke($invocation);
 
         $this->assertEquals('advice', $result, "Advice should change the return value of invocation");
@@ -39,10 +39,9 @@ class MethodAroundInterceptorTest extends AbstractMethodInterceptorTest
         };
         $invocation = $this->getInvocation($sequence);
 
-        $interceptor = new MethodAroundInterceptor($advice);
+        $interceptor = new AroundInterceptor($advice);
         $result = $interceptor->invoke($invocation);
         $this->assertEquals('invocation', $result, "Advice should return an original return value");
         $this->assertEquals(array('advice', 'invocation', 'advice'), $sequence, "Around logic should work");
     }
 }
-
