+++ tests/Go/Aop/Framework/BeforeInterceptorTest.php
@@ -10,7 +10,7 @@
 
 namespace Go\Aop\Framework;
 
-class MethodBeforeInterceptorTest extends AbstractMethodInterceptorTest
+class BeforeInterceptorTest extends AbstractMethodInterceptorTest
 {
     public function testAdviceIsCalledBeforeInvocation()
     {
@@ -18,11 +18,10 @@ class MethodBeforeInterceptorTest extends AbstractMethodInterceptorTest
         $advice     = $this->getAdvice($sequence);
         $invocation = $this->getInvocation($sequence);
 
-        $interceptor = new MethodBeforeInterceptor($advice);
+        $interceptor = new BeforeInterceptor($advice);
         $result = $interceptor->invoke($invocation);
 
         $this->assertEquals('invocation', $result, "Advice should not affect the return value of invocation");
         $this->assertEquals(array('advice', 'invocation'), $sequence, "Before advice should be invoked before invocation");
     }
 }
-
