+++ tests/Go/Aop/Framework/ClosureStaticMethodInvocationTest.php
@@ -92,7 +92,7 @@ class ClosureStaticMethodInvocationTest extends \PHPUnit_Framework_TestCase
     {
         $child  = $this->getMock(self::FIRST_CLASS_NAME, array('none'));
         $value  = 'test';
-        $advice = new MethodBeforeInterceptor(function () use (&$value) {
+        $advice = new BeforeInterceptor(function () use (&$value) {
             $value = 'ok';
         });
